+++ Grapple/src/Grapple/Renderer/Buffer.h
@@ -18,7 +18,7 @@ namespace Grapple
 		Float4,
 	};
 
-	constexpr size_t ShaderDataTypeSize(ShaderDataType dataType)
+	constexpr uint32_t ShaderDataTypeSize(ShaderDataType dataType)
 	{
 		switch (dataType)
 		{
